ALLY-465 Relocate CSV rendering to outputs.go
