+++ cli/cmd/outputs_test.go
